\name{Cattle}
\docType{data}
\alias{Cattle}
\title{Information on Genotyped Cattle}
\description{
 Breed name and year of birth of the cattle whose genotypes are included in files \link{Chr1.phased}, and \link{Chr2.phased}.
}
\usage{data(Cattle)}
\format{Data frame containing information on genotyped cattle. The columns contain the ID of the individual (\code{Indiv}), the year of birth (\code{Born}), and the breed name (\code{Breed}).
}
\keyword{datasets}