\name{ExamplePed}
\docType{data}
\alias{ExamplePed}
\title{Pedigree of Hinterwald Cattle}
\description{
  This data set gives a small subset of the pedigree of Hinterwald cattle suitable for demonstration purposes.
}
\usage{ExamplePed}
\format{A data frame with columns for individual ID, sire ID, dam ID, sex, breed, year of birth, and a simulated breeding value}
\keyword{datasets}