\name{help.opticont4mb}
\Rdversion{1.1}
\alias{help.opticont4mb}
\title{Displays Available Objective Functions and Constraints for Function \code{opticont4mb}.
}
\description{
Displays available objective functions and constraints for function \link{opticont4mb}.
}
\usage{
help.opticont4mb(K, phen)}

\arguments{
\item{K}{Named list containing one or more kinship matrices.}
\item{phen}{Data frame with N rows and various columns, with animal IDs in column 1 and sex in column 2. The other columns may contain breeding values or migrant contributions.}
}

\details{
Lists available objective functions and constraints for function \link{opticont4mb}.
}

\examples{
#see ?opticont4mb
}

\author{Robin Wellmann}
