\name{Kin}
\docType{data}
\alias{Kin}
\title{Simulated phenotypes of Hinterwald cattle}
\description{
 A list of class kinMatrices containing kinship matrix fA and the matrices fB and fN for the animals included in datat set \link{Phen}. Matrices fB and fN  are needed to compute conditional kinship fD. 
}
\usage{Kin}
\format{A list of class kinMatrices.}
\keyword{datasets}