\name{map}
\docType{data}
\alias{map}
\title{Marker Map for Cattle}
\description{
 Marker map for SNPs from cattle chromosomes 1 - 2  (only the first parts of the chromosomes). The corresponding genotypes are included in \link{Chr1.phased} and  \link{Chr2.phased}.
}
\usage{data(map)}
\format{Data frame containing the marker map including marker name (\code{Name}), chromosome number (\code{Chr}), position in base pairs (\code{Position}), position in centiMorgan (\code{cM}), and position in mega base pairs (\code{Mb}).
}
\keyword{datasets}