\name{PedigWithErrors}
\docType{data}
\alias{PedigWithErrors}
\title{Pedigree of Hinterwald cattle}
\description{
  This data set gives the pedigree of Hinterwald cattle with some artificially introduced errors.
}
\usage{PedigWithErrors}
\format{A data frame with columns \code{Indiv} (individual ID), \code{Sire}, \code{Dam}, \code{Sex}, \code{Breed}, and column \code{Born} with year of birth.}
\keyword{datasets}