\name{Cattle}
\docType{data}
\alias{Cattle}
\title{Phenotypes of Genotyped Cattle}
\description{
 Simulated phenotypes of cattle whose genotypes are included in files \link{Chr1.phased}, and \link{Chr2.phased}.
}
\usage{data(Cattle)}
\format{Data frame containing information on genotyped cattle.
The columns contain the ID of the individual (\code{Indiv}), the year of birth (\code{Born}), the breed name (\code{Breed}), a breeding value (\code{BV}), the sex (\code{Sex}), and the herd (\code{herd}).
}
\keyword{datasets}