\name{ExamplePed}
\docType{data}
\alias{ExamplePed}
\title{Pedigree of Hinterwald Cattle}
\description{
  This data set gives a small subset of the pedigree of Hinterwald cattle suitable for demonstration purposes.
}
\usage{data(ExamplePed)}
\format{Data frame with columns \code{Indiv} (individual ID), \code{Sire}, \code{Dam}, \code{Sex}, \code{Breed}, \code{Born} with year of birth, and simulated breeding value \code{BV}.}
\keyword{datasets}