\name{help.opticont}
\Rdversion{1.1}
\alias{help.opticont}
\title{Displays Available Objective Functions and Constraints for Function \code{opticont}.
}
\description{
Displays available objective functions and constraints for function \link{opticont}.
}
\usage{
help.opticont(K, phen)}

\arguments{
\item{K}{Named list containing one or more NxN Kinship matrices for the N breeding animals.}
\item{phen}{Data frame with N rows and various columns, with animal IDs in column 1 and sex in column 2. The other columns may contain breeding values or migrant contributions.}
}

\details{
Lists available objective functions and constraints for function \link{opticont}.
}

\examples{
#see ?opticont
}

\author{Robin Wellmann}
