\name{PedigWithErrors}
\docType{data}
\alias{PedigWithErrors}
\title{Pedigree of Hinterwald cattle}
\description{
  This data set gives the pedigree of Hinterwald cattle with some artificially introduced errors and simulated breeding values.
}
\usage{data(PedigWithErrors)}
\format{A data frame with columns \code{Indiv} (individual ID), \code{Sire}, \code{Dam}, \code{Sex}, \code{Breed}, \code{Born} with year of birth, and column \code{BV} with simulated breeding values.}
\keyword{datasets}