\name{PedigWithErrors}
\docType{data}
\alias{PedigWithErrors}
\title{Pedigree of Hinterwald cattle}
\description{
  This data set gives the pedigree of Hinterwald cattle with some artificially introduced errors.
}
\usage{PedigWithErrors}
\format{A data frame with columns for individual ID, sire ID, dam ID, sex, breed, and year of birth}
\keyword{datasets}