\name{Phen}
\docType{data}
\alias{Phen}
\title{Simulated Phenotypes of Hinterwald Cattle}
\description{
 A data frame simulated breeding values of some Hinterwald cattle
 with offspring born in 2006 or 2007.
}
\usage{data(Phen)}
\format{ A data frame with individual IDs (\code{Indiv}), sexes (\code{Sex}), breeding values (\code{BV}), and native contribution (\code{NC}).}
\keyword{datasets}