\name{Phen}
\docType{data}
\alias{Phen}
\title{Simulated phenotypes of Hinterwald cattle}
\description{
 A data frame with migrant contributions and simulated breeding values of some Hinterwald cattle
 with offspring born in 2006 or 2007.
}
\usage{Phen}
\format{ A data frame with columns for individual ID (Indiv), sex, breeding values (BV), and migrant contributions (MC).}
\keyword{datasets}