\name{Phen}
\docType{data}
\alias{Phen}
\title{Simulated Phenotypes of Hinterwald Cattle}
\description{
 A data frame simulated breeding values of some Hinterwald cattle
 with offspring born in 2006 or 2007.
}
\usage{Phen}
\format{ A data frame with columns for individual ID (Indiv), sex, and breeding values (BV).}
\keyword{datasets}